+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 ## [Unreleased][]
 
+## [Release][3.0.1] [3.0.1][3.0.0-3.0.1] - 2024-02-06
+
 ### Added
 
   - [#147](https://github.com/GDATASoftwareAG/robotframework-flaui/issues/147) New Keywords
@@ -574,6 +576,10 @@ General bugfixing from keywords and improvements from artifact building and test
 [github.io]: https://gdatasoftwareag.github.io/robotframework-flaui
 
 [unreleased]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/3.0.0...main
+
+[3.0.1]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/3.0.1
+
+[3.0.0-3.0.1]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/3.0.0...3.0.1
 
 [3.0.0]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/3.0.0